+++ examples/rider-plugin/src/main/java/org/intellij/examples/rider/plugin/HelloAction.java
@@ -3,17 +3,32 @@ package org.intellij.examples.rider.plugin;
 import com.intellij.icons.AllIcons;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.ui.Messages;
+import com.intellij.openapi.actionSystem.PlatformDataKeys;
+import com.intellij.openapi.ui.DialogWrapper;
+import org.jetbrains.annotations.Nullable;
+
+import javax.swing.*;
 
 public class HelloAction extends AnAction {
-  @Override
-  public void actionPerformed(AnActionEvent e) {
-    Messages.showInfoMessage("Hello World!", "Hello");
-  }
+    @Override
+    public void actionPerformed(AnActionEvent e) {
+        DialogWrapper dialogWrapper = new DialogWrapper(PlatformDataKeys.PROJECT.getData(e.getDataContext())) {
+            {
+                init();
+            }
 
-  @Override
-  public void update(AnActionEvent e) {
-    super.update(e);
-    e.getPresentation().setIcon(AllIcons.Ide.Info_notifications);
-  }
+            @Nullable
+            @Override
+            protected JComponent createCenterPanel() {
+                return new HelloForm().getPanel();
+            }
+        };
+        dialogWrapper.showAndGet();
+    }
+
+    @Override
+    public void update(AnActionEvent e) {
+        super.update(e);
+        e.getPresentation().setIcon(AllIcons.Ide.Info_notifications);
+    }
 }