+++ examples/rider-plugin/src/main/java/org/intellij/examples/rider/plugin/HelloForm.java
@@ -1,0 +1,11 @@
+package org.intellij.examples.rider.plugin;
+
+import javax.swing.*;
+
+public class HelloForm {
+    private JPanel myPanel;
+
+    JPanel getPanel() {
+        return myPanel;
+    }
+}